refactor(router): scope null object helper to routers
